Extract owner and location side effect modules for routing/location
As with our service registry, we need these to live in separate files
so they can be copied over without change and so be able to perform
imports while *also* performing a module merge.
