+++ packages/@ember/routing/location.ts
@@ -123,9 +123,3 @@ export type UpdateCallback = (url: string) => void;
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface Registry extends Record<string, Location | undefined> {}
-
-declare module '@ember/owner' {
-  export interface DIRegistry {
-    location: Registry;
-  }
-}
