+++ packages/@ember/routing/none-location.ts
@@ -134,9 +134,3 @@ NoneLocation.reopen({
   path: '',
   rootURL: '/',
 });
-
-declare module '@ember/routing/location' {
-  export interface Registry {
-    none: NoneLocation;
-  }
-}
